requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openai==1.88.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sentencepiece==0.2.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
textblob==0.19.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
